# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "lending-bank-gateway",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/loans/p2p-disbursements": {
      "post": {
        "tags": [
          "loans"
        ],
        "summary": "P2P Disbursement",
        "operationId": "p2p_disbursement_api_v1_loans_p2p_disbursements_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/P2PDisbursementRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitAckEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/loans/p2p-repayments": {
      "post": {
        "tags": [
          "loans"
        ],
        "summary": "P2P Repayment",
        "operationId": "p2p_repayment_api_v1_loans_p2p_repayments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/P2PRepaymentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepaymentAckEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/loans/p2p-repayments/{biz_seq_no}/status": {
      "get": {
        "tags": [
          "loans"
        ],
        "summary": "Query Repayment Status",
        "description": "还款专用状态查询（对接文档 v0.6.1 §4.2）：本地 order 状态 + wedap 逐笔明细。\n\n为什么不复用 `/bank-funds/status`（5.5 通用查询）：5.5 自 v0.6.0 起虽能查还款**顶层\n状态**，但 `debtSettled`（Lending 核销债务的唯一依据）与 `steps[]`（逐笔与银行对账）\n**只有本专用接口提供**。受理响应为 PROCESSING 时按本接口轮询至终态。\n\n本地 order 不存在 → 404 且不打 wedap：存在性判据在本方台账，不拿外部系统当权威\n（亦防 bizSeqNo 探测）。wedap 不可用 → 降级 `unavailable`，本地 orderStatus 仍可读。",
        "operationId": "query_repayment_status_api_v1_loans_p2p_repayments__biz_seq_no__status_get",
        "parameters": [
          {
            "name": "biz_seq_no",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Biz Seq No"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepaymentStatusEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/bank-funds/status": {
      "get": {
        "tags": [
          "bank-funds"
        ],
        "summary": "Query Status",
        "description": "查询本地 order 状态 + wedap 实时状态（wedap 不可用时降级为 unavailable=True）。",
        "operationId": "query_status_api_v1_bank_funds_status_get",
        "parameters": [
          {
            "name": "bizSeqNo",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bizseqno"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Query Status Api V1 Bank Funds Status Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DisbursementInfo": {
        "properties": {
          "txnAmount": {
            "type": "string",
            "title": "Txnamount"
          },
          "currencyCode": {
            "type": "string",
            "title": "Currencycode"
          },
          "userId": {
            "type": "string",
            "title": "Userid",
            "default": ""
          },
          "userName": {
            "type": "string",
            "title": "Username",
            "default": ""
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "txnAmount",
          "currencyCode"
        ],
        "title": "DisbursementInfo"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MoneyWriteOperationStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "RECONCILING",
          "SUCCEEDED",
          "REJECTED"
        ],
        "title": "MoneyWriteOperationStatus",
        "description": "v2.2 §8.2 ``operationStatus`` 受控枚举；终态 SUCCEEDED/REJECTED 后不得回处理中。"
      },
      "MoneyWriteOutcome": {
        "type": "string",
        "enum": [
          "NOT_APPLIED",
          "PENDING",
          "UNKNOWN",
          "ACCEPTED"
        ],
        "title": "MoneyWriteOutcome",
        "description": "v2.2 §8.2 ``outcome`` 封闭枚举：同步成功之外的业务意图事实。"
      },
      "MoneyWriteRetryPolicy": {
        "type": "string",
        "enum": [
          "NEVER",
          "RETRY_SAME_KEY_AFTER",
          "POLL_STATUS",
          "REAUTH_AND_REPLAY",
          "CORRECT_AND_NEW_INTENT"
        ],
        "title": "MoneyWriteRetryPolicy",
        "description": "v2.2 §8.2 ``retryPolicy``：写客户端唯一可执行的重试策略词表。"
      },
      "P2PDisbursementRequest": {
        "properties": {
          "bizSeqNo": {
            "type": "string",
            "title": "Bizseqno"
          },
          "channelId": {
            "type": "string",
            "title": "Channelid",
            "default": ""
          },
          "transType": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Transtype"
          },
          "disbursementInfo": {
            "$ref": "#/components/schemas/DisbursementInfo"
          },
          "lenders": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lenders"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "bizSeqNo",
          "transType",
          "disbursementInfo"
        ],
        "title": "P2PDisbursementRequest"
      },
      "P2PRepaymentRequest": {
        "properties": {
          "bizSeqNo": {
            "type": "string",
            "title": "Bizseqno"
          },
          "transType": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Transtype"
          },
          "loanNo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Loanno",
            "description": "借据单号。wedap 侧自 2026-07-23 起无条件必填（loan_repayment_txn.loan_no NOT NULL，灰度开关同迁移移除）；缺失时本网关返 400 GW_400_VALIDATION，不下发 wedap。schema 上标 optional 仅为统一错误形态，业务上必传。"
          },
          "repaymentInfo": {
            "$ref": "#/components/schemas/RepaymentInfo"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "bizSeqNo",
          "transType",
          "repaymentInfo"
        ],
        "title": "P2PRepaymentRequest"
      },
...
      "RepaymentAck": {
        "properties": {
          "bizSeqNo": {
            "type": "string",
            "title": "Bizseqno"
          },
          "txnStatus": {
            "type": "string",
            "title": "Txnstatus"
          },
          "orderStatus": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Orderstatus"
          },
          "errorCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errorcode"
          },
          "errorMsg": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errormsg"
          },
          "inFlight": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inflight"
          },
          "outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MoneyWriteOutcome"
              },
              {
                "type": "null"
              }
            ],
            "description": "业务意图事实；**同步完成的成功响应省略本字段**（由成功模型表达）。PENDING/UNKNOWN/ACCEPTED 一律不得重提，须查 statusUrl 或进对账。"
          },
          "operationStatus": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MoneyWriteOperationStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "retryPolicy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MoneyWriteRetryPolicy"
              },
              {
                "type": "null"
              }
            ],
            "description": "写客户端唯一可执行的重试策略。POLL_STATUS=查 statusUrl；CORRECT_AND_NEW_INTENT=修正请求并换新 bizSeqNo（bizSeqNo 即幂等键，同键重发只会拿回同一条冻结响应）；NEVER=不得重试。"
          },
          "resubmitAllowed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resubmitallowed",
            "description": "是否允许用**同一幂等键**重提。本网关无 9000 在册的权威放行策略，故恒为 false（fail-closed）。"
          },
          "operationId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operationid",
            "description": "durable operation 标识（= bizSeqNo）。与 statusUrl 成对出现。"
          },
          "statusUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Statusurl",
            "description": "本 operation 的查单地址，受同 tenant/principal 的对象级授权。"
          },
          "debtSettled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Debtsettled"
          },
          "globalTxId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Globaltxid"
          },
          "detailStatus": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detailstatus"
          }
        },
        "type": "object",
        "required": [
          "bizSeqNo",
          "txnStatus"
        ],
        "title": "RepaymentAck",
        "description": "还款受理响应 data 段（对接文档 v0.6.1 §4.2，SubmitAck + DTC 三字段）。\n\n这三个字段是 v0.5.0 随组合交易引擎新增的，**必须在 openapi 里可见**——否则调用方\n无从知道 ``debtSettled`` 存在，只能靠口头传达（上游至今未接该字段正是此成因）。\n\n- ``debtSettled``：契约指定的**债务核销唯一依据**，仅全部资金步骤成功时为 true\n- ``globalTxId``：wedap 组合交易实例号，报障/对账时提供给 wedap\n- ``detailStatus``：细分状态，供排查参考；**状态机仍以 txnStatus 为准**\n逐笔 ``steps[]`` 与 ``strandedAmount`` 不在受理响应，须查本模块的专用状态查询端点。"
      },
      "RepaymentAckEnvelope": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "$ref": "#/components/schemas/RepaymentAck"
          },
          "error": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          }
        },
        "type": "object",
        "required": [
          "success",
          "data",
          "error",
          "trace_id"
        ],
        "title": "RepaymentAckEnvelope",
        "description": "还款受理 200 响应统一 envelope。"
      },
      "RepaymentInfo": {
        "properties": {
          "txnAmount": {
            "type": "string",
            "title": "Txnamount"
          },
          "currencyCode": {
            "type": "string",
            "title": "Currencycode"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "txnAmount",
          "currencyCode"
        ],
        "title": "RepaymentInfo"
      },
      "RepaymentStatusData": {
        "properties": {
          "bizSeqNo": {
            "type": "string",
            "title": "Bizseqno"
          },
          "orderStatus": {
            "type": "string",
            "title": "Orderstatus"
          },
          "wedap": {
            "additionalProperties": true,
            "type": "object",
            "title": "Wedap"
          }
        },
        "type": "object",
        "required": [
          "bizSeqNo",
          "orderStatus",
          "wedap"
        ],
        "title": "RepaymentStatusData",
        "description": "还款专用状态查询 200 响应 data 段。\n\n``wedap`` 是**薄透传**段：正常时为 wedap 专用状态查询响应（含 ``debtSettled`` /\n``strandedAmount`` / ``steps[]``），wedap 不可达时降级为\n``{\"unavailable\": true, \"reason\": ...}``。逐笔 ``steps[]`` 的结构由 wedap 契约决定、\n本网关不复刻，故此段保持开放对象——但顶层三键必须在 openapi 里可见，否则调用方\n连「响应长什么样」都要靠口头传达（2026-08-10 独立评审 finding）。"
      },
      "RepaymentStatusEnvelope": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "$ref": "#/components/schemas/RepaymentStatusData"
          },
          "error": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          }
        },
        "type": "object",
        "required": [
          "success",
          "data",
          "error",
          "trace_id"
        ],
        "title": "RepaymentStatusEnvelope",
        "description": "还款专用状态查询 200 响应统一 envelope。"
      },
...
      "SubmitAck": {
        "properties": {
          "bizSeqNo": {
            "type": "string",
            "title": "Bizseqno"
          },
          "txnStatus": {
            "type": "string",
            "title": "Txnstatus"
          },
          "orderStatus": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Orderstatus"
          },
          "errorCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errorcode"
          },
          "errorMsg": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errormsg"
          },
          "inFlight": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inflight"
          },
          "outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MoneyWriteOutcome"
              },
              {
                "type": "null"
              }
            ],
            "description": "业务意图事实；**同步完成的成功响应省略本字段**（由成功模型表达）。PENDING/UNKNOWN/ACCEPTED 一律不得重提，须查 statusUrl 或进对账。"
          },
          "operationStatus": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MoneyWriteOperationStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "retryPolicy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MoneyWriteRetryPolicy"
              },
              {
                "type": "null"
              }
            ],
            "description": "写客户端唯一可执行的重试策略。POLL_STATUS=查 statusUrl；CORRECT_AND_NEW_INTENT=修正请求并换新 bizSeqNo（bizSeqNo 即幂等键，同键重发只会拿回同一条冻结响应）；NEVER=不得重试。"
          },
          "resubmitAllowed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resubmitallowed",
            "description": "是否允许用**同一幂等键**重提。本网关无 9000 在册的权威放行策略，故恒为 false（fail-closed）。"
          },
          "operationId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operationid",
            "description": "durable operation 标识（= bizSeqNo）。与 statusUrl 成对出现。"
          },
          "statusUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Statusurl",
            "description": "本 operation 的查单地址，受同 tenant/principal 的对象级授权。"
          }
        },
        "type": "object",
        "required": [
          "bizSeqNo",
          "txnStatus"
        ],
        "title": "SubmitAck",
        "description": "写原语提交响应 data 段最小字段契约（collect/distribute/refund/reversal 共用）。\n\n- ``txnStatus``：wedap 同步语义（SUCCESS 已同步终态 / PROCESSING 异步在途 /\n  FAILED / RESULT_UNKNOWN）\n- ``orderStatus``：gateway 台账 :class:`~app.domain.states.OrderStatus`\n  （ACCEPTED/SUBMITTED/SUCCEEDED/FAILED/RESULT_UNKNOWN/...），受理 vs 落账\n  以此字段区分；权威终态仍以 ``GET /bank-funds/status`` 为准。\n  缺失的两种情形：契约上线前受理单的幂等重放（first_response 冻结不回填）；\n  in-flight 重放（``inFlight=true``，零查询路径不读 order 行）——均应查单。\n- 未设置字段序列化省略（端点挂 ``response_model_exclude_unset``）：线格式对既有\n  调用方纯增量（不出现 ``errorCode: null`` 噪声）。不能用 model_serializer 丢 None\n  ——wrap serializer 的 dict 返回注解会把 openapi schema 打回 additionalProperties。"
      },
      "SubmitAckEnvelope": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "$ref": "#/components/schemas/SubmitAck"
          },
          "error": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          }
        },
        "type": "object",
        "required": [
          "success",
          "data",
          "error",
          "trace_id"
        ],
        "title": "SubmitAckEnvelope",
        "description": "写原语提交 200 响应统一 envelope（:func:`app.core.envelope.ok` 的类型化形态）。"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}